backend/main.py
from fastapi import FastAPI, WebSocket, WebSocketDisconnect, HTTPException, Depends, UploadFile, File
from fastapi.middleware.cors import CORSMiddleware
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
import asyncio
import json
import uuid
from datetime import datetime
from typing import Dict, List, Optional
import redis
import asyncpg
from sqlalchemy import create_engine, MetaData, Table, Column, String, DateTime, Text, Boolean, Integer
from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
import jwt
import bcrypt
import openai
import anthropic
import google.generativeai as genai
from voice_engine import VoiceEngine
from notion_integration import NotionIntegrator
from ai_orchestrator import AIOrchestrator

app = FastAPI(title="NEXUS PRO API", version="1.0.0")
security = HTTPBearer()

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Database setup
DATABASE_URL = "postgresql+asyncpg://user:password@localhost/nexuspro"
engine = create_async_engine(DATABASE_URL)
redis_client = redis.Redis(host='localhost', port=6379, decode_responses=True)

# WebSocket connections
connections: Dict[str, WebSocket] = {}
rooms: Dict[str, List[str]] = {}

# Initialize services
voice_engine = VoiceEngine()
notion_integrator = NotionIntegrator()
ai_orchestrator = AIOrchestrator()

class ConnectionManager:
    def __init__(self):
        self.active_connections: List[WebSocket] = []
        self.room_connections: Dict[str, List[WebSocket]] = {}

    async def connect(self, websocket: WebSocket, room_id: str = "default"):
        await websocket.accept()
        self.active_connections.append(websocket)
        if room_id not in self.room_connections:
            self.room_connections[room_id] = []
        self.room_connections[room_id].append(websocket)

    def disconnect(self, websocket: WebSocket, room_id: str = "default"):
        self.active_connections.remove(websocket)
        if room_id in self.room_connections:
            self.room_connections[room_id].remove(websocket)

    async def broadcast_to_room(self, message: str, room_id: str):
        if room_id in self.room_connections:
            for connection in self.room_connections[room_id]:
                try:
                    await connection.send_text(message)
                except:
                    pass

manager = ConnectionManager()

# Authentication
def verify_token(credentials: HTTPAuthorizationCredentials = Depends(security)):
    try:
        payload = jwt.decode(credentials.credentials, "secret", algorithms=["HS256"])
        return payload["user_id"]
    except:
        raise HTTPException(status_code=401, detail="Invalid token")

# Voice Coding Endpoints
@app.post("/api/voice/transcribe")
async def transcribe_audio(audio: UploadFile = File(...), user_id: str = Depends(verify_token)):
    result = await voice_engine.transcribe_audio(audio.file.read())
    return {"transcription": result["text"], "confidence": result["confidence"]}

@app.post("/api/voice/execute")
async def execute_voice_command(command: dict, user_id: str = Depends(verify_token)):
    result = await voice_engine.execute_command(command["text"], user_id)
    await manager.broadcast_to_room(json.dumps({
        "type": "voice_command",
        "user_id": user_id,
        "command": command["text"],
        "result": result
    }), user_id)
    return result

# AI Orchestration Endpoints
@app.post("/api/ai/orchestrate")
async def orchestrate_ai(request: dict, user_id: str = Depends(verify_token)):
    result = await ai_orchestrator.process_request(request["prompt"], request.get("context", {}))
    return result

@app.get("/api/ai/models/status")
async def get_ai_models_status():
    return await ai_orchestrator.get_models_status()

# Collaboration Endpoints
@app.websocket("/ws/collaboration/{room_id}")
async def websocket_collaboration(websocket: WebSocket, room_id: str):
    await manager.connect(websocket, room_id)
    try:
        while True:
            data = await websocket.receive_text()
            message = json.loads(data)
            
            # Process different message types
            if message["type"] == "canvas_update":
                await manager.broadcast_to_room(data, room_id)
            elif message["type"] == "cursor_move":
                await manager.broadcast_to_room(data, room_id)
            elif message["type"] == "code_change":
                await manager.broadcast_to_room(data, room_id)
                
    except WebSocketDisconnect:
        manager.disconnect(websocket, room_id)

# Notion Integration Endpoints
@app.post("/api/notion/sync")
async def sync_notion(request: dict, user_id: str = Depends(verify_token)):
    result = await notion_integrator.sync_workspace(request["workspace_id"], user_id)
    return result

@app.post("/api/notion/create-task")
async def create_notion_task(task: dict, user_id: str = Depends(verify_token)):
    result = await notion_integrator.create_task(task, user_id)
    return result

# Deployment Endpoints
@app.post("/api/deploy/vercel")
async def deploy_to_vercel(project: dict, user_id: str = Depends(verify_token)):
    # Implement Vercel deployment
    deployment_id = str(uuid.uuid4())
    redis_client.set(f"deployment:{deployment_id}", json.dumps({
        "status": "deploying",
        "platform": "vercel",
        "user_id": user_id,
        "created_at": datetime.now().isoformat()
    }))
    
    # Simulate deployment process
    asyncio.create_task(simulate_deployment(deployment_id, project))
    return {"deployment_id": deployment_id, "status": "initiated"}

@app.post("/api/deploy/netlify")
async def deploy_to_netlify(project: dict, user_id: str = Depends(verify_token)):
    deployment_id = str(uuid.uuid4())
    redis_client.set(f"deployment:{deployment_id}", json.dumps({
        "status": "deploying",
        "platform": "netlify",
        "user_id": user_id,
        "created_at": datetime.now().isoformat()
    }))
    
    asyncio.create_task(simulate_deployment(deployment_id, project))
    return {"deployment_id": deployment_id, "status": "initiated"}

@app.get("/api/deploy/status/{deployment_id}")
async def get_deployment_status(deployment_id: str):
    status = redis_client.get(f"deployment:{deployment_id}")
    if status:
        return json.loads(status)
    raise HTTPException(status_code=404, detail="Deployment not found")

# Database Endpoints
@app.post("/api/database/connect")
async def connect_database(config: dict, user_id: str = Depends(verify_token)):
    connection_id = str(uuid.uuid4())
    redis_client.set(f"db_connection:{connection_id}", json.dumps({
        "type": config["type"],
        "host": config["host"],
        "database": config["database"],
        "user_id": user_id,
        "connected_at": datetime.now().isoformat()
    }))
    return {"connection_id": connection_id, "status": "connected"}

@app.post("/api/database/query")
async def execute_query(request: dict, user_id: str = Depends(verify_token)):
    # Simulate query execution
    return {
        "results": [{"id": 1, "name": "Sample Data"}],
        "execution_time": "0.05s",
        "rows_affected": 1
    }

# Game Generation Endpoints
@app.post("/api/games/generate")
async def generate_game(prompt: dict, user_id: str = Depends(verify_token)):
    game_id = str(uuid.uuid4())
    
    # Generate game based on prompt
    game_data = {
        "id": game_id,
        "name": prompt.get("name", "Generated Game"),
        "type": prompt.get("type", "3d"),
        "code": generate_game_code(prompt["description"]),
        "assets": generate_game_assets(prompt["description"]),
        "created_at": datetime.now().isoformat()
    }
    
    redis_client.set(f"game:{game_id}", json.dumps(game_data))
    return game_data

# Mobile App Generation Endpoints
@app.post("/api/mobile/generate")
async def generate_mobile_app(prompt: dict, user_id: str = Depends(verify_token)):
    app_id = str(uuid.uuid4())
    
    # Generate mobile app
    app_data = {
        "id": app_id,
        "name": prompt.get("name", "Generated App"),
        "platforms": prompt.get("platforms", ["ios", "android"]),
        "files": generate_mobile_files(prompt["description"]),
        "dependencies": generate_dependencies(prompt["description"]),
        "created_at": datetime.now().isoformat()
    }
    
    redis_client.set(f"mobile_app:{app_id}", json.dumps(app_data))
    return app_data

# Analytics Endpoints
@app.get("/api/analytics/dashboard")
async def get_analytics_dashboard(user_id: str = Depends(verify_token)):
    return {
        "performance": {
            "response_time": "120ms",
            "uptime": "99.9%",
            "requests_per_minute": 1250
        },
        "security": {
            "vulnerabilities": 0,
            "security_score": 98,
            "last_scan": datetime.now().isoformat()
        },
        "usage": {
            "active_users": 1500,
            "projects_created": 450,
            "deployments": 230
        }
    }

# Marketplace Endpoints
@app.get("/api/marketplace/plugins")
async def get_plugins():
    return [
        {
            "id": "plugin-1",
            "name": "Advanced Code Formatter",
            "description": "Professional code formatting with custom rules",
            "category": "Development",
            "price": 29.99,
            "rating": 4.8,
            "downloads": 15000
        },
        {
            "id": "plugin-2", 
            "name": "AI Code Reviewer",
            "description": "Automated code review with AI suggestions",
            "category": "AI Tools",
            "price": 49.99,
            "rating": 4.9,
            "downloads": 8500
        }
    ]

@app.post("/api/marketplace/install/{plugin_id}")
async def install_plugin(plugin_id: str, user_id: str = Depends(verify_token)):
    redis_client.sadd(f"user_plugins:{user_id}", plugin_id)
    return {"status": "installed", "plugin_id": plugin_id}

# Helper functions
async def simulate_deployment(deployment_id: str, project: dict):
    await asyncio.sleep(5)  # Simulate deployment time
    
    redis_client.set(f"deployment:{deployment_id}", json.dumps({
        "status": "completed",
        "url": f"https://{project['name']}-{deployment_id[:8]}.vercel.app",
        "completed_at": datetime.now().isoformat()
    }))

def generate_game_code(description: str) -> str:
    return f"""
// Generated game code for: {description}
const scene = new THREE.Scene();
const camera = new THREE.PerspectiveCamera(75, window.innerWidth / window.innerHeight, 0.1, 1000);
const renderer = new THREE.WebGLRenderer();

// Game logic here
function animate() {{
    requestAnimationFrame(animate);
    renderer.render(scene, camera);
}}
animate();
"""

def generate_game_assets(description: str) -> dict:
    return {
        "textures": ["ground.jpg", "sky.jpg"],
        "models": ["player.obj", "environment.obj"],
        "sounds": ["background.mp3", "effects.wav"]
    }

def generate_mobile_files(description: str) -> dict:
    return {
        "App.tsx": "// React Native App",
        "package.json": "{}",
        "android/build.gradle": "// Android config"
    }

def generate_dependencies(description: str) -> list:
    return ["react-native", "@react-navigation/native", "react-redux"]

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)