backend/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.21.1
websockets==11.0.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==1.10.7
sqlalchemy==2.0.7
asyncpg==0.27.0
redis==4.5.4
celery==5.2.7
aiofiles==23.1.0
httpx==0.24.0
openai==0.27.2
anthropic==0.2.6
google-generativeai==0.1.0
assemblyai==0.15.0
notion-client==2.0.0
firebase-admin==6.1.0
supabase==1.0.3
psycopg2-binary==2.9.5
pymongo==4.3.3
motor==3.1.2
aiomysql==0.1.1
aioredis==2.0.1
boto3==1.26.101
google-cloud-storage==2.8.0
azure-storage-blob==12.14.1
stripe==5.4.0
twilio==8.1.0
sendgrid==6.10.0
pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.2
pandas==2.0.0
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
tensorflow==2.12.0
torch==2.0.0
transformers==4.28.1
langchain==0.0.154
chromadb==0.3.21
pinecone-client==2.2.1
weaviate-client==3.15.5
pytest==7.2.2
pytest-asyncio==0.21.0
black==23.3.0
flake8==6.0.0
mypy==1.2.0
pre-commit==3.2.2
docker==6.0.1
kubernetes==26.1.0
prometheus-client==0.16.0
grafana-api==1.0.3
sentry-sdk[fastapi]==1.19.1
loguru==0.7.0
structlog==23.1.0
pyjwt==2.6.0
cryptography==40.0.1
bcrypt==4.0.1
argon2-cffi==21.3.0
email-validator==2.0.0
phonenumbers==8.13.7
pycountry==22.3.5
babel==2.12.1
pytz==2023.3
dateutil==2.8.2
cachetools==5.3.0
tenacity==8.2.2
backoff==2.2.1
ratelimit==2.2.1
slowapi==0.1.8
starlette-prometheus==0.9.0
prometheus-fastapi-instrumentator==5.11.2